Move bundle viewsets to own modules
